apigenerator/resources/1 - Project/1 - BaseProject/Project/requirements.txt
chardet
enforce-typing
flask
flask-swagger-ui
marshmallow-sqlalchemy
pyyaml
regex
requests
sqlalchemy
ujson